auth_service/requirements.txt
redis==4.4.2
python-dotenv==0.21.0
fastapi==0.110.0
orjson==3.8.6
pydantic==2.6.3
pydantic-settings==2.2.1
uvicorn[standard]==0.12.2
gunicorn==21.2.0
backoff==2.2.1
psycopg2_binary==2.9.9
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
asyncpg==0.29.0
passlib==1.7.4
StrEnum==0.4.15
pycryptodomex==3.17
pybase64==1.3.2
PyJWT==2.8.0
python-multipart==0.0.9
sqlalchemy==2.0.29
async-fastapi-jwt-auth==0.6.4
flake8==7.0.0
click==7.1.2
alembic==1.13.1
